refactor: organize component tests
