+++ pkg/agentcompose/coverage_shape_workflows_test.go
@@ -23,16 +23,7 @@ func TestIntegrationWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testServiceProxyRoutesRedirectAndProxy(t)
 	TestServiceProxyRoutesUseGuestHostTarget(t)
 	testServiceEnsureProxyReadyStartPaths(t)
-	TestJupyterTargetReachableUsesGuestHostTarget(t)
-	TestSessionDriverStartSessionVMSavesRuntimeProxyState(t)
 	testServiceStreamingAPIs(t)
-	testRuntimeProviderSelectsConfiguredRuntime(t)
-	testLLMClientGenerateHandlesSuccessAndFailures(t)
-	testWebhookHandlerStoresEvent(t)
-	testEventQueryHandlers(t)
-	testPrepareFileWorkspaceCopiesContent(t)
-	testPrepareGitWorkspaceClonesRootAndTarget(t)
-	testRegisterWorkspaceRoutesUploadAndList(t)
 	testSupportConstructorsAndHelpers(t)
 	testSupportControlPlaneStartAndConfigHelpers(t)
 	testSupportSetupRegistersServiceGraph(t)
@@ -42,14 +33,11 @@ func TestIntegrationWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testServiceReconcilePersistedSessionsMarksStalePendingFailed(t)
 	TestServiceAndBridgeReconcileMicrosandboxRuntimeTypeBranches(t)
 	testServiceProtoConversionHelpers(t)
-	testAgentRunSummariesScansAllSessions(t)
 	testAgentDefinitionConfigStoreCRUDAndWorkspaceProtection(t)
 	testLoaderCreateBindsAgentDefinitionProvider(t)
 	testAgentDefinitionValidationAndProtoMapping(t)
 	testAgentDefinitionCreateSession(t)
 	testDeleteAgentDefinitionStopsSessionsAndKeepsDeletedInList(t)
-	testDashboardOverviewAggregatorCountsRuns(t)
-	testDashboardOverviewHubWatchInitialAndNotify(t)
 	testRuntimeLLMFacadeCoverageWorkflows(t)
 	testWebhookIntegrationEventDispatchRunsMatchingLoader(t)
 }
@@ -75,16 +63,7 @@ func TestE2EWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testServiceProxyRoutesRedirectAndProxy(t)
 	TestServiceProxyRoutesUseGuestHostTarget(t)
 	testServiceEnsureProxyReadyStartPaths(t)
-	TestJupyterTargetReachableUsesGuestHostTarget(t)
-	TestSessionDriverStartSessionVMSavesRuntimeProxyState(t)
 	testServiceStreamingAPIs(t)
-	testRuntimeProviderSelectsConfiguredRuntime(t)
-	testLLMClientGenerateHandlesSuccessAndFailures(t)
-	testWebhookHandlerStoresEvent(t)
-	testEventQueryHandlers(t)
-	testPrepareFileWorkspaceCopiesContent(t)
-	testPrepareGitWorkspaceClonesRootAndTarget(t)
-	testRegisterWorkspaceRoutesUploadAndList(t)
 	testSupportConstructorsAndHelpers(t)
 	testSupportControlPlaneStartAndConfigHelpers(t)
 	testSupportSetupRegistersServiceGraph(t)
@@ -94,14 +73,11 @@ func TestE2EWebhookWorkspaceAndLoaderWorkflow(t *testing.T) {
 	testServiceReconcilePersistedSessionsMarksStalePendingFailed(t)
 	TestServiceAndBridgeReconcileMicrosandboxRuntimeTypeBranches(t)
 	testServiceProtoConversionHelpers(t)
-	testAgentRunSummariesScansAllSessions(t)
 	testAgentDefinitionConfigStoreCRUDAndWorkspaceProtection(t)
 	testLoaderCreateBindsAgentDefinitionProvider(t)
 	testAgentDefinitionValidationAndProtoMapping(t)
 	testAgentDefinitionCreateSession(t)
 	testDeleteAgentDefinitionStopsSessionsAndKeepsDeletedInList(t)
-	testDashboardOverviewAggregatorCountsRuns(t)
-	testDashboardOverviewHubWatchInitialAndNotify(t)
 	testRuntimeLLMFacadeCoverageWorkflows(t)
 	testServiceReconcilePersistedSessionsMarksStaleProjectRunsFailed(t)
 	testWebhookIntegrationEventDispatchRunsMatchingLoader(t)
@@ -124,10 +100,6 @@ func testRuntimeLLMFacadeCoverageWorkflows(t *testing.T) {
 	TestEnsureSessionOpenCodeAnthropicUsesFacadeEnv(t)
 	TestEnsureSessionLLMFacadeConfigBootstrapsFromSessionEnvProvider(t)
 	TestCodexFacadeCanUseAnthropicOnlySessionEnvProvider(t)
-	TestLLMClientGenerateChatCompletions(t)
-	TestLLMClientGenerateChatCompletionsPlainText(t)
-	TestLLMClientGenerateSendsOutputSchema(t)
-	TestLLMClientResolveProtocol(t)
 	TestServiceGenerateLLMChatCompletionsProtocol(t)
 	testRuntimeLLMFacadeRoundTrips(t)
 }
